+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.  
 This project adheres to [Semantic Versioning](https://semver.org).
 
-## [2.0.6] - Unreleased
+## [2.1.0] - 2022-01-03
 
 ### Added